Cnc4 is troubled.  Refactored HiddenState5 into a record.  Check for performance regression.
Specifically Cnc4 should fork N-1 workers because the finalize step
itself runs alongside the workers (and will issue a replacement if it
blocks on a get).

But when I try this "correct" approach I see the same kind of problems
on embarassingly_par that I've seen in the past with other schedulers
-- spurious serialization.  The second work unit isn't picked up until
the first is completed.