Added all existing tests on the spec language to tests_spec. Here are the findings:
Failures in supposedly good ones (i.e. the ones that are accepted by the old translator):

1. The "prescription" relation is bi-directional. (test004.cnc)
2. Duplicate collection declarations are allowed. (test014.cnc)
3. Syntax for tag components is not handled. (test016.cnc and many other failures)
4. ENV is a keyword (same as env). (test033.cnc)
5. <tag> -> env should be allowed. (testnq_cnc.cnc, testsearchinsortedtree.cnc)

Failures in supposedly bad ones (i.e. the ones that are not accepted by the old translator):

1. Multi-line comment enclosed in /*.... */. (test005.cnc)
2. <- env -> (test034.cnc)
3. -> env <- (test035.cnc)
4. env, env (test036.cnc)
5. Items produced only by env cannot be consumed only by env (test045.cnc)

I think the problems in the first group should be fixed for backward compatibility. As for the second group, the first 4 are actually not an issue (rather, they are improvements that don't break backward compatibility). The last one does need to be addressed, though.

One other thing I observed is that the new translator reports only the first error it encounters. If that's the case, then I guess the syntax for tag components should be addressed first so that those tests that currently fail because of that can be tested for real (and not masked by the tag component syntax error).